Add Lun Functionality

+++ BLL/EGITService.cs
@@ -254,31 +254,31 @@ namespace BLL
         }
         public GenerateErrorDto AddLun(CreateLunDto lun)
         {
-            StorageDto linkedStorage = GetStorage(lun.StorageID);
 
             LunDto newLun = new LunDto
             {
                 LunName = lun.LunName,
-                LunTotalRAM = lun.LunTotalRAM,
-                StorageID = lun.StorageID,
-                LunRemainingRAM=lun.LunTotalRAM,
+                LunTotalSpace = lun.LunTotalSpace,
+                LunRemainingSpace=lun.LunTotalSpace,
+                StorageID = lun.StorageID
 
             };
+            StorageDto linkedStorage = GetStorage(lun.StorageID);
+
+            if (linkedStorage == null)
+            {
+                return new GenerateErrorDto { Response = "Storage Not Found, Cannot Create Lun!" };
+            }
+
+            if (lun.LunTotalSpace > linkedStorage.StorageRemainingSpace)
+            {
+                return new GenerateErrorDto { Response = "Lun RAM cannot Exceed Storage RAM!", IsValid = false };
+            }
             try
             {
-                if (lun.LunTotalRAM > linkedStorage.StorageRemainingRAM)
-                {
-                    return new GenerateErrorDto { Response = "Lun RAM cannot Exceed Storage RAM!", IsValid = false };
-                }
-                else
-                {
-
-                    EGITRepository.AddLun(mapper.Map<Lun>(newLun));
-                    CalculateStorageRAM(newLun.StorageID);
-                    return new GenerateErrorDto { Response = "Lun Added Successfully!", IsValid = true };
-
-                }
-
+                EGITRepository.AddLun(mapper.Map<Lun>(newLun));
+                CalculateStorageSpace(newLun.StorageID);
+                return new GenerateErrorDto { Response = "Lun Added Successfully!", IsValid = true };
             }
 
             catch (Exception)
@@ -301,12 +301,11 @@ namespace BLL
             {
                 return new GenerateErrorDto { Response = "Lun Not Found, Cannot Delete This Lun!", IsValid = false };
             }
-
-            LunDto LunToBeDeleted = GetLun(LunID);
+          
             try
             {
                 EGITRepository.DeleteLun(LunID);
-                CalculateStorageRAM(LunToBeDeleted.StorageID);
+                CalculateStorageSpace(lunToBeDeleted.StorageID);
                 return new GenerateErrorDto { Response = "Lun Deleted Successfully!", IsValid = true };
             }
 
@@ -320,42 +319,52 @@ namespace BLL
         {
 
             LunDto LunToBeUpdated = GetLun(LunID);
-            StorageDto linkedStorage = GetStorage(LunToBeUpdated.StorageID);
-            var StorageRemainingRAM = -1;
+            StorageDto oldlinkedStorage = GetStorage(LunToBeUpdated.StorageID);
+            StorageDto newlinkedStorage = GetStorage(UpdatedLun.StorageID);
+            var remainingStorage = 0;
 
-
-
-            if (LunToBeUpdated != null && linkedStorage!=null)
+            if (LunToBeUpdated == null)
             {
-                StorageRemainingRAM = linkedStorage.StorageRemainingRAM + LunToBeUpdated.LunTotalRAM;
-                var newLunRemaining = (UpdatedLun.LunTotalRAM - LunToBeUpdated.LunTotalRAM) + LunToBeUpdated.LunRemainingRAM;
+                return new GenerateErrorDto { Response = "Lun Not Found, Cannot Update Lun!", IsValid = false };
+            }
 
-                LunToBeUpdated.LunName = UpdatedLun.LunName;
-                LunToBeUpdated.LunTotalRAM = UpdatedLun.LunTotalRAM;
-                LunToBeUpdated.StorageID = UpdatedLun.StorageID;
-                LunToBeUpdated.LunRemainingRAM = newLunRemaining;
-            }  
 
+            if (newlinkedStorage == null || oldlinkedStorage == null)
+            {
+                return new GenerateErrorDto { Response = "Storage Not Found, Cannot Update Lun!", IsValid = false };
+
+            }
+          
+            if(LunToBeUpdated.StorageID == UpdatedLun.StorageID)
+            {
+                remainingStorage =  oldlinkedStorage.StorageRemainingSpace + LunToBeUpdated.LunTotalSpace;
+            }
+            else
+            {
+                oldlinkedStorage.StorageRemainingSpace += LunToBeUpdated.LunTotalSpace;
+                EGITRepository.UpdateStorage(mapper.Map<Storage>(oldlinkedStorage));
+                remainingStorage = newlinkedStorage.StorageRemainingSpace;
+
+            }
+
+            LunToBeUpdated.LunName = UpdatedLun.LunName;
+            LunToBeUpdated.LunTotalSpace = UpdatedLun.LunTotalSpace;
+            LunToBeUpdated.StorageID = UpdatedLun.StorageID;
+            if (LunToBeUpdated.LunTotalSpace > remainingStorage)
+            {
+                return new GenerateErrorDto { Response = "Lun RAM cannot Exceed Storage RAM!", IsValid = false };
+
+            }
             try
             {
-                if (LunToBeUpdated.LunTotalRAM> StorageRemainingRAM)
-                {
-                    return new GenerateErrorDto { Response = "Lun RAM cannot Exceed Storage RAM!", IsValid = false };
-
-                }
-
-                else
-                {
-
-                    EGITRepository.UpdateLun(mapper.Map<Lun>(LunToBeUpdated));
-                    CalculateStorageRAM(LunToBeUpdated.StorageID);
-                    return new GenerateErrorDto { Response = "Lun Updated Successfully!", IsValid = true };
-                }
+                EGITRepository.UpdateLun(mapper.Map<Lun>(LunToBeUpdated));
+                CalculateStorageSpace(LunToBeUpdated.StorageID);
+                CalculateLunSpace(LunToBeUpdated.LunID);
+                return new GenerateErrorDto { Response = "Lun Updated Successfully!", IsValid = true };
             }
             catch
             {
                 return new GenerateErrorDto { Response = "Error Updating The Lun!", IsValid = false };
-
             }
 
         }
@@ -377,8 +386,8 @@ namespace BLL
             {
                 StorageName = storage.StorageName,
                 StorageType = storage.StorageType,
-                StorageTotalRAM = storage.StorageTotalRAM,
-                StorageRemainingRAM = storage.StorageTotalRAM
+                StorageTotalSpace = storage.StorageTotalSpace,
+                StorageRemainingSpace = storage.StorageTotalSpace
 
             };
 
@@ -396,6 +405,10 @@ namespace BLL
         }
         public GenerateErrorDto DeleteStorage(int StorageID)
         {
+            if(GetStorage(StorageID)== null)
+            {
+                return new GenerateErrorDto { Response = "Storage Not Found, Cannot Delete Storage!", IsValid = false };
+            }
             try
             {
                 EGITRepository.DeleteStorage(StorageID);
@@ -415,18 +428,25 @@ namespace BLL
 
             if (StorageToBeUpdated != null)
             {
-                var newRemaining = (UpdatedStorage.StorageTotalRAM - StorageToBeUpdated.StorageTotalRAM) + StorageToBeUpdated.StorageRemainingRAM;
 
                 StorageToBeUpdated.StorageName = UpdatedStorage.StorageName;
                 StorageToBeUpdated.StorageType = UpdatedStorage.StorageType;
-                StorageToBeUpdated.StorageTotalRAM = UpdatedStorage.StorageTotalRAM;
-                StorageToBeUpdated.StorageRemainingRAM = newRemaining;
+                StorageToBeUpdated.StorageTotalSpace = UpdatedStorage.StorageTotalSpace;
 
-                EGITRepository.UpdateStorage(mapper.Map<Storage>(StorageToBeUpdated));
-                return new GenerateErrorDto { Response = "Storage Updated Successfully!", IsValid = true };
+            }
+            else
+            {
+                return new GenerateErrorDto { Response = "Storage Not Found, Cannot Update Storage!", IsValid = false };
+
             }
 
-            else
+            try
+            {
+                EGITRepository.UpdateStorage(mapper.Map<Storage>(StorageToBeUpdated));
+                CalculateStorageSpace(StorageToBeUpdated.StorageID);
+                return new GenerateErrorDto { Response = "Storage Updated Successfully!", IsValid = true };
+            }
+            catch
             {
                 return new GenerateErrorDto { Response = "Error Updating The Storage!", IsValid = false };
             }
@@ -461,24 +481,64 @@ namespace BLL
                 return new GenerateErrorDto { Response = "Cluster Not Found!", IsValid = false };
             }
         }
-        public GenerateErrorDto CalculateStorageRAM(int StorageID)
+        public GenerateErrorDto CalculateStorageSpace(int StorageID)
         {
             StorageDto storage = GetStorage(StorageID);
-            if (storage != null)
+            if (storage == null)
             {
-                List<Lun> luns = EGITRepository.GetStorageLuns(StorageID);
-                var RemainingSum  = storage.StorageTotalRAM - luns.Sum(l => l.LunTotalRAM);
-                storage.StorageRemainingRAM = RemainingSum;
+                return new GenerateErrorDto { Response = "Storage Not Found, Cannot Calculate RAM!", IsValid = false };
 
+            }
+            List<Lun> luns = EGITRepository.GetStorageLuns(StorageID);
+
+            if (luns == null)
+            {
+               return new GenerateErrorDto { Response = "Storage Doesn't Contain Luns, Cannot Calculate RAM!", IsValid = false };
+
+            }
+            var RemainingSum = storage.StorageTotalSpace - luns.Sum(l => l.LunTotalSpace);
+            storage.StorageRemainingSpace = RemainingSum;
+
+            try
+            {
                 EGITRepository.UpdateStorage(mapper.Map<Storage>(storage));
                 return new GenerateErrorDto { Response = " RAM Calculated Successfully!", IsValid = true };
             }
-
-        else
+             catch
             {
                 return new GenerateErrorDto { Response = "Error Calculating RAM!", IsValid = false };
             }
             
+        }
+
+        public GenerateErrorDto CalculateLunSpace(int LunID)
+        {
+            LunDto lun = GetLun(LunID);
+            if (lun == null)
+            {
+                return new GenerateErrorDto { Response = "Lun Not Found, Cannot Calculate Lun RAM!", IsValid = false };
+
+            }
+            List<VM> vms = EGITRepository.GetLunVMs(LunID);
+            if (vms == null)
+            {
+                return new GenerateErrorDto { Response = "Lun Contains no Vms, Cannot Calculate Lun RAM!", IsValid = false };
+
+            }
+            var RemainingSum = lun.LunTotalSpace - vms.Sum(l => l.RAM);
+            lun.LunRemainingSpace = RemainingSum;
+
+            try
+            {
+                EGITRepository.UpdateLun(mapper.Map<Lun>(lun));
+                return new GenerateErrorDto { Response = "  Lun RAM Calculated Successfully!", IsValid = true };
+
+            }
+            catch
+            {
+                return new GenerateErrorDto { Response = "Error Calculating Lun RAM!", IsValid = false };
+            }
+
         }
 
         public List<VMDto> GetAllVMs()
@@ -534,6 +594,7 @@ namespace BLL
                 {
                     EGITRepository.AddVM(mapper.Map<VM>(newVM));
                     this.CalculateNodeRemainingSpace(newVM.NodeID);
+                    this.CalculateLunSpace(newVM.LunID);
                     return new GenerateErrorDto { Response = "VM Added Successfully!", IsValid = true };
                 }
                 
@@ -601,6 +662,7 @@ namespace BLL
                 {
                     EGITRepository.UpdateVM(mapper.Map<VM>(newVM));
                     this.CalculateNodeRemainingSpace(newVM.NodeID);
+                    this.CalculateLunSpace(newVM.LunID);
                     return new GenerateErrorDto { Response = "VM Updated Successfully!", IsValid = true };
                 }
             }
@@ -624,6 +686,7 @@ namespace BLL
             {
                 EGITRepository.DeleteVM(VMID);
                 this.CalculateNodeRemainingSpace(VMToBeDeleted.NodeID);
+                this.CalculateLunSpace(VMToBeDeleted.LunID);
                 return new GenerateErrorDto { Response = "VM Deleted Successfully!", IsValid = true };
             }
